+++ examples/package.json
@@ -28,7 +28,6 @@
     "@chakra-ui/react": "2.5.5",
     "framer-motion": "^10.11.5",
     "@createform/react": "workspace:*",
-    "@createform/nativeform": "workspace:*",
     "react": "18.2.0",
     "react-dom": "18.2.0"
   },
