--- packages/nativeform/jest.config.js
@@ -1,9 +1,0 @@
-module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'jsdom',
-  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
-  moduleNameMapper: {
-    '\\.(css|less)$': 'identity-obj-proxy'
-  },
-  setupFilesAfterEnv: ['@testing-library/jest-dom/extend-expect']
-};
