--- packages/nativeform/package.json
@@ -1,67 +1,0 @@
-{
-  "name": "@createform/nativeform",
-  "version": "3.3.29",
-  "description": "The ReactJS form library.",
-  "keywords": [
-    "react",
-    "hooks",
-    "form",
-    "forms",
-    "form-validation",
-    "validation",
-    "typescript",
-    "react-hooks"
-  ],
-  "author": "Jucian0 <juciano@outlook.com.br>",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/jucian0/useform",
-    "homepage": "https://useform.org/"
-  },
-  "dependencies": {
-    "@createform/object-utils": "workspace:*",
-    "@createform/validation": "workspace:*"
-  },
-  "devDependencies": {
-    "tsup": "^6.7.0",
-    "@createform/tsconfig": "workspace:*",
-    "jest": "^29.5.0",
-    "ts-jest": "^29.0.5",
-    "jest-each": "^29.5.0",
-    "@jest/globals": "29.5.0",
-    "yup": "^1.0.2",
-    "@types/yup": "^0.29.13",
-    "@types/react": "18.0.31",
-    "zod": "^3.21.4",
-    "@faker-js/faker": "^7.4.0",
-    "typescript": "^4.9.3",
-    "@testing-library/react": "^13.4.0",
-    "@types/jest": "29.5.0",
-    "@testing-library/jest-dom": "5.16.5",
-    "jest-environment-jsdom": "29.5.0",
-    "react": "^18.0.0",
-    "react-dom": "18.2.0"
-  },
-  "peerDependencies": {
-    "react": ">=18.0.0"
-  },
-  "main": "./dist/index.js",
-  "module": "./dist/index.mjs",
-  "types": "./dist/index.d.ts",
-  "sideEffects": false,
-  "license": "MIT",
-  "files": [
-    "dist/**"
-  ],
-  "scripts": {
-    "build": "tsup src/index.ts --format esm,cjs --dts --external react",
-    "dev": "tsup src/index.ts --format esm,cjs --watch --dts --external react",
-    "lint": "eslint \"src/**/*.ts*\"",
-    "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist",
-    "test": "jest",
-    "test:watch": "jest --watch"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
