+++ packages/react/src/CreateForm/index.ts
@@ -1,2 +1,2 @@
 export * from "./CreateForm";
-export * from "./UseForm";
+export * from "./Types";
